+++ aws_profile_manager/cli.py
@@ -30,9 +30,13 @@ class MyCLI(click.MultiCommand):
     def get_command(self, ctx, name):
         ns = {}
         fn = os.path.join(plugin_folder, name + '.py')
-        with open(fn) as f:
-            code = compile(f.read(), fn, 'exec')
-            eval(code, ns, ns)
+        try:
+            with open(fn) as f:
+                code = compile(f.read(), fn, 'exec')
+                eval(code, ns, ns)
+        except FileNotFoundError:
+            logging.error(f'No such command {name}')
+            return
         return ns['cli']
 
 class RootFlags(object):