+++ aws_profile_manager/commands/profile-add.py
@@ -1,4 +1,3 @@
-import sys
 import click
 import logging
 from aws_profile_manager import Common, Add