+++ aws_profile_manager/commands/profile-list.py
@@ -1,5 +1,4 @@
 import click
-import logging
 from aws_profile_manager import Common
 
 @click.command()