+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.1](https://github.com/99stealth/aws-profile-manager/tree/v0.5.1) - 2022-07-05
+### Added
+- Error handling in case there are no such command
+### Changed
+- Strings formatings from `"{}".format(var_name)` to `f"{var_name}"` style
+
 ## [0.5.0](https://github.com/99stealth/aws-profile-manager/tree/v0.5.0) - 2022-06-14
 ### Added
 - Add `profile-list` command that simply shows all profiles